git-svn-id: https://svn.code.sf.net/p/pianobooster/code/trunk@133 ba081f5d-443b-49a7-ac4b-446c3f91f371

+++ PianoBooster/src/Tempo.h
@@ -50,7 +50,7 @@ public:
     void reset()
     {
         // 120 beats per minute is the default
-        setMidiTempo(( 60 * MICRO_SECOND ) / 120 );
+        setMidiTempo(static_cast<int>(( 60 * MICRO_SECOND ) / 120 ));
         m_jumpAheadDelta = 0;
     }
